Bring pipeline instruments into line with the rest of the instrument init code.
